backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
newspaper3k>=0.2.8
lxml_html_clean>=0.1.0
sentence-transformers>=3.0.0
numpy>=1.26.0
